feat: Implement side navigation drawer in dashboard
The code changes add a new file `side_navigation_drawer.dart` which contains the implementation of a side navigation drawer for the dashboard. The side navigation drawer includes menu items for drivers, users, and trips pages. It also includes a header with an image. The `SideNavigationDrawer` widget is used in the `MainApp` widget as the initial route.

This commit addresses the need to have a navigation menu in the dashboard and improves the user experience by providing easy access to different pages.