feat: Update dependencies and add Firestore methods
This commit updates the dependencies in the `pubspec.yaml` file to include the latest versions of `cloud_firestore`, `firebase_auth`, `firebase_core`, and `firebase_database`. It also adds a new file `firestore_methods.dart` which contains the implementation of Firestore methods for retrieving drivers data. These changes enhance the functionality of the application by integrating Firebase services and providing access to drivers data.